+++ test/Reflex/Test/HostSpec.hs
@@ -17,13 +17,31 @@ import           Reflex.Test.Host
 
 import           Control.Monad            (forM_)
 import           Control.Monad.IO.Class   (liftIO)
+import           Data.Functor
 import qualified Data.List                as L
 import           Data.These
 
 
--- TODO add a more complicated test in here
--- TODO be sure to test PostBuildT stuff
+-- | network that ensures postbuild event was triggered
+postbuild_network
+  :: forall t m
+   . (t ~ SpiderTimeline Global, m ~ SpiderHost Global)
+  => (AppIn t () () -> TestGuestMonad t m (AppOut t Bool ()))
+postbuild_network AppIn {..} = do
+  pbev            <- getPostBuild
+  didPBTriggerBeh <- hold False (pbev $> True)
+  return AppOut { _appOut_behavior = didPBTriggerBeh, _appOut_event = never }
 
+test_postbuild :: Test
+test_postbuild = TestLabel "postbuild" $ TestCase $ runSpiderHost $ do
+  appFrame <- getAppFrame postbuild_network ()
+  -- tick the appFrame once which will give us the output behavior value of the previous frame (which triggered the postbuild event)
+  out      <- tickAppFrame appFrame (Just (That ()))
+  liftIO $ out @?= [(True, Nothing)]
+
+
+-- | Basic network that flips returns a behavior that is the input behavior flipped
+-- and an event that is the input behavior's value added to the input events value.
 basic_network
   :: forall t m
    . (t ~ SpiderTimeline Global, m ~ SpiderHost Global)
@@ -46,3 +64,4 @@ test_basic = TestLabel "basic" $ TestCase $ runSpiderHost $ do
 spec :: Spec
 spec = describe "Reflex.Test.App" $ do
   fromHUnitTest test_basic
+  fromHUnitTest test_postbuild