+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## 0.1.1.0
 - generalized reflex constraints
 - added `PostBuildT` support
+- fixed a crash that would happen if there were no event subscribers
 
 ## 0.1.0.0
 - first release